requirements.txt
load_dotenv
langchain
openai
duckduckgo-search
google-cloud-aiplatform